Move test to itest package
